// src/app/model/saida-detalhe-api.ts
// Generated by https://quicktype.io

import { CartaoCredito } from "./cartao-credito";
import { ParcelaApi } from "./parcela-api";

export interface saidaDetalhesApi {
    id:            number;
    obs:           string;
    nome:          string;
    meioPagto:     string;
    parcelas:      ParcelaApi[];
    total:         number;
    diaVencimento: number;
    cartao:        CartaoCredito;
    xdeParcelas:   string;
}